+++ index.js
@@ -8,6 +8,7 @@ import 'react-mdl/extra/material.js';
 
 import MaterialMixin from './src/mixins/MaterialMixin';
 import PerformanceMixin from './src/mixins/PerformanceMixin';
+import ScrollingMixin from './src/mixins/ScrollingMixin';
 
 // GUI Elements
 
@@ -92,6 +93,7 @@ import Version from './src/elements/Version/Version';
 export default {
     MaterialMixin,
     PerformanceMixin,
+    ScrollingMixin,
 
     Alert,
     BreadcrumbList,